pyproject.toml
[project]
name = "forum-api"
version = "0.0.2"
description = "Forum API web app for TelerikAcademy"
authors = [
    {name = "fcgith"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = "3.12.5"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "cffi==1.17.1",
    "click==8.1.8",
    "cryptography==44.0.2",
    "fastapi==0.115.12",
    "h11==0.14.0",
    "idna==3.10",
    "jinja2==3.1.6",
    "markupsafe==3.0.2",
    "pycparser==2.22",
    "pydantic==2.11.2",
    "pydantic-core==2.33.1",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "starlette==0.46.1",
    "typing-inspection==0.4.0",
    "typing-extensions==4.13.1",
    "uvicorn==0.34.0",
    "pyjwt>=2.10.1,<3.0.0",
    "psycopg2-binary==2.9.10",
    "python-dotenv==1.1.0",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"